pyproject.toml
[tool.poetry]
name = "reddit"
version = "0.1.0"
description = ""
authors = ["Nicolas Vincent"]
license=""
readme="README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.14.0"
praw = "^7.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
